Rename events tests
