+++ tsconfig.json
@@ -76,6 +76,10 @@
 		"./tests/unit/aspect.ts",
 		"./tests/unit/async/iteration.ts",
 		"./tests/unit/async/timing.ts",
+		"./tests/unit/events/all.ts",
+		"./tests/unit/events/browserOnly.ts",
+		"./tests/unit/events/common.ts",
+		"./tests/unit/events/nodeOnly.ts",
 		"./tests/unit/global.ts",
 		"./tests/unit/has.ts",
 		"./tests/unit/lang.ts",
@@ -85,10 +89,6 @@
 		"./tests/unit/number.ts",
 		"./tests/unit/object.ts",
 		"./tests/unit/observers/ObjectObserver.ts",
-		"./tests/unit/on/all.ts",
-		"./tests/unit/on/browserOnly.ts",
-		"./tests/unit/on/common.ts",
-		"./tests/unit/on/nodeOnly.ts",
 		"./tests/unit/queue.ts",
 		"./tests/unit/streams/ByteLengthQueuingStrategy.ts",
 		"./tests/unit/streams/CountQueuingStrategy.ts",
